Add ApplicationComponent for bootstraping.
